Refactor: MinioRepository PR 리뷰 반영
